requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
bleach==6.0.0
boto3==1.28.30
botocore==1.31.30
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20230430
Django==4.2.4
django-admin-sortable2==2.1.9
django-appconf==1.0.5
django-author==1.2.0
django-channels==0.7.0
django-cors==0.1
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-imagekit==4.1.0
django-import-export==3.2.0
django-import-export-celery==1.6
django-indexer==0.3.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-paging==0.2.5
django-phonenumber-field==7.1.0
django-redis==5.3.0
django-s3direct==2.0.3
django-storages==1.13.2
django-summernote==0.8.20.0
django-templated-mail==1.1.1
django-templatetag-sugar==1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.4.2
docutils==0.20.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.3
fcm-django==2.0.0
FeinCMS==23.8.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gotrue==1.0.3
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
h11==0.14.0
html2text==2020.1.16
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.1
lockfile==0.12.2
MarkupPy==1.14
msgpack==1.0.5
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
phonenumbers==8.13.18
pilkit==2.0
Pillow==10.0.0
postgrest==0.10.8
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.1
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.8.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
raven==6.10.0
realtime==1.0.0
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
sentry-sdk==1.29.2
setuptools-git==1.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
South==1.0.2
sqlparse==0.4.4
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
xlrd==2.0.1
xlwt==1.3.0
